+++ src/propertystore.hpp
@@ -1,170 +1,7 @@
 #ifndef PROPERTYSTORE_HPP
 #define PROPERTYSTORE_HPP
 
-#ifdef DEBUG_PROPERTY
-#include "utility/debug.hpp"
-#endif
-
 #include "property.hpp"
-
-
-#if defined(ARDUINO_ARCH_AVR) && defined(E2END)
-#include <avr/eeprom.h>
-#include <avr/io.h>
-
-// Keep KEY_NAME_MAX_SIZE equal to NVS_KEY_NAME_MAX_SIZE from nvs.h esp32
-#define KEY_NAME_MAX_SIZE 16        // Includes null-terminator.
-const size_t mem_segment_size = max(max(IntegerProperty::size, BooleanProperty::size), RealProperty::size);
-const size_t mem_entry_size   = KEY_NAME_MAX_SIZE + mem_segment_size;
-const size_t mem_max_entries  = round((E2END + 1) / mem_entry_size);
-
-class Preferences
-{
-    public:
-        Preferences();
-        bool begin(const char* mem_namespace = nullptr);
-        void end();
-        bool isKey(const char* key);
-        size_t getBytes(const char* key, void* value, size_t len);
-        size_t putBytes(const char* key, const void* value, size_t len);
-        bool remove(const char* key);
-        // bool wipeMemory();
-    private:
-        size_t n_entries;
-        size_t getEntry(const char* key, uint8_t* buf = nullptr);
-};
-
-Preferences::Preferences()
-{
-
-}
-
-bool Preferences::begin(const char* mem_namespace)
-{
-    if(!eeprom_is_ready()) return false;
-
-    // Count entries in memory
-    uint8_t block[mem_entry_size] = {0};
-    while(n_entries < mem_max_entries)
-    {
-        #ifdef WIPE_MEMORY
-        eeprom_write_block(block, (uint8_t*)(n_entries * mem_entry_size), mem_entry_size);
-        #else
-        eeprom_read_block(block, (uint8_t*)(n_entries * mem_entry_size), mem_entry_size);
-        if(block[0] == 0) break;
-        #endif
-        n_entries++;
-    }
-    #ifdef WIPE_MEMORY
-    n_entries = 0;
-    #endif
-    return true;
-}
-
-void Preferences::end()
-{
-
-}
-
-/* Find the block of memory in eeprom associated with the provided key, and store it in buf. */
-size_t Preferences::getEntry(const char* key, uint8_t* buf)
-{
-    size_t i = 0;
-    uint8_t block[mem_entry_size] = {0};
-    if(buf == nullptr) buf = block;
-
-    while(i < n_entries)
-    {
-        eeprom_read_block(buf, (void*)(i * mem_entry_size), mem_entry_size);
-        if(strncmp(key, (const char*)buf, strlen(key)) == 0) return i;
-        i++;
-    }
-    return i;
-}
-
-bool Preferences::isKey(const char* key)
-{
-    return getEntry(key) < n_entries;
-}
-
-size_t Preferences::putBytes(const char* key, const void* value, size_t len)
-{
-    if(len != mem_segment_size)
-    {
-        return 0;
-    }
-    uint8_t block[mem_entry_size] = {0};
-    size_t i = getEntry(key, block);
-    if(i < n_entries)
-    {
-        // Update existing entry
-        memcpy(block + KEY_NAME_MAX_SIZE, value, len);
-        eeprom_update_block(block + KEY_NAME_MAX_SIZE, (void*)(i * mem_entry_size + KEY_NAME_MAX_SIZE), mem_segment_size);
-        return len;
-    }
-    else if(n_entries < mem_max_entries)
-    {
-        // Create new entry
-        size_t keylength = strlen(key);
-        if(keylength >= KEY_NAME_MAX_SIZE) return 0;
-        memcpy(block, key, keylength);
-        memcpy(block + KEY_NAME_MAX_SIZE, value, len);
-        eeprom_write_block(block, (void*)(i * mem_entry_size), mem_entry_size);
-        n_entries++;
-        return len;
-    }
-    else
-    {
-        return 0;
-    }
-    
-}
-
-size_t Preferences::getBytes(const char* key, void* value, size_t len)
-{
-    if(len != mem_segment_size) return 0;
-
-    uint8_t block[mem_entry_size] = {0};
-    size_t i = getEntry(key, block);
-    if(i < n_entries)
-    {
-        memcpy(value, block + KEY_NAME_MAX_SIZE, len);
-        return len;
-    }
-    return 0;
-}
-
-bool Preferences::remove(const char* key)
-{
-    size_t i = getEntry(key);
-    if(i < n_entries)
-    {
-        uint8_t block[mem_entry_size] = {0};
-        // Update existing entry
-        eeprom_write_block(block, (void*)(i * mem_entry_size), mem_entry_size);
-        return true;
-    }
-    else return false;
-}
-
-/* bool Preferences::wipeMemory()
-{
-    uint8_t block[16] = {0};
-    for(size_t i = 0; i < (E2END + 1); i += 16)
-    {
-        eeprom_write_block(block, (void*)i, 16);
-    }
-    return true;
-} */
-
-#elif defined(ARDUINO_ARCH_ESP32)
-#include "Print.h"
-#include "Preferences.h"
-const size_t mem_segment_size = std::max({IntegerProperty::size, BooleanProperty::size, RealProperty::size});
-#endif
-
-#if defined(ARDUINO_ARCH_ESP32) || defined(ARDUINO_ARCH_AVR)
-
 class BaseStore
 {
     protected:
@@ -172,6 +9,33 @@ class BaseStore
     public:
         const size_t size;
         virtual BaseProperty* get_property(uint32_t i) = 0;
+
+        struct Iterator
+        {
+            public:
+                using dtype = BaseProperty*;
+                using Pointer = dtype*;
+                using Reference = dtype&;
+                Iterator(Pointer ptr) : m_ptr(ptr) {}    
+            
+                Reference operator*() const { return *m_ptr; }
+                Pointer operator->() { return m_ptr; }
+
+                // Prefix increment
+                Iterator& operator++() { m_ptr++; return *this; }  
+
+                // Postfix increment
+                Iterator operator++(int) { Iterator tmp = *this; ++(*this); return tmp; }
+
+                friend bool operator== (const Iterator& a, const Iterator& b) { return a.m_ptr == b.m_ptr; };
+                friend bool operator!= (const Iterator& a, const Iterator& b) { return a.m_ptr != b.m_ptr; };
+
+            private:
+                Pointer m_ptr;
+        };
+        virtual Iterator begin() = 0;
+        virtual Iterator end() = 0;
+        
 };
 
 template<size_t SIZE>
@@ -179,51 +43,28 @@ class PropertyStore: public BaseStore
 {
     private:
         BaseProperty* properties[SIZE];
-        Preferences memory;
-        const char* memory_namespace = "properties";
-        bool memory_opened;
-
-        void save_to_memory(BaseProperty* p)
-        {
-            #ifdef DEBUG_PROPERTY
-            PRINT("[Properties]: Saving ", p->get_name(), " to memory.");
-            #endif
-            if(!memory_opened)
-            {
-                #ifdef DEBUG_PROPERTY
-                PRINT("[Properties]: Opening memory");
-                #endif
-                memory_opened = memory.begin(memory_namespace);
-                if(!memory_opened)
-                {    
-                    #ifdef DEBUG_PROPERTY
-                    PRINT("[Properties]: Memory failed to open");
-                    #endif
-                    return;
-                }
-            }
-            uint8_t mem[mem_segment_size] = {0};
-            p->save_to_bytes(mem);
-            memory.putBytes(p->get_name(), mem, mem_segment_size);
-            p->saved();
-        }
         
     public:
-        // PropertyStore(std::initializer_list<BaseProperty*> props):
-        //     properties(props)
-        // {}
         PropertyStore():
-            BaseStore(SIZE),
-            memory_opened(false)
+            BaseStore(SIZE)
         {}
 
         PropertyStore(BaseProperty* const (&props)[SIZE]):
-            BaseStore(SIZE),
-            memory_opened(false)
+            BaseStore(SIZE)
         {
             assign_properties(props);
         }
 
+        Iterator begin()
+        {
+            return Iterator(&properties[0]);
+        }
+
+        Iterator end()
+        {
+            return Iterator(&properties[SIZE]);
+        }
+        
         PropertyStore<SIZE> operator=(const PropertyStore<SIZE>& rhs)
         {
             assign_properties(rhs.properties);
@@ -237,104 +78,6 @@ class PropertyStore: public BaseStore
                 properties[i] = props[i];
             }
         }
-        
-        /// @brief WARNING: `PropertyStore` should be constructed with a property array 
-        /// or `assign_properties()` should have been called before `begin()`
-        bool begin()
-        {
-            #ifdef DEBUG_PROPERTY
-            PRINT("[Properties]: Opening memory");
-            #endif
-            memory_opened = memory.begin(memory_namespace);
-            
-            if(!memory_opened) 
-            {
-                #ifdef DEBUG_PROPERTY
-                PRINT("[Properties]: Memory failed to open");
-                #endif
-                return false;
-            }
-
-            // Loop through all registered properties and see if they already have a 
-            // value stored in memory. If not, save its current (default) value.
-            for(BaseProperty* p : properties)
-            {
-                if(memory.isKey(p->get_name()))
-                {
-                    // Load existing value from memory
-                    #ifdef DEBUG_PROPERTY
-                    PRINT("[Properties]: Loading ", p->get_name(), " from memory.");
-                    #endif
-                    uint8_t mem[mem_segment_size] = {0};
-                    size_t len = memory.getBytes(p->get_name(), mem, mem_segment_size);
-                    if(len != mem_segment_size)
-                    {
-                        #ifdef DEBUG_PROPERTY
-                        PRINT("[Properties]: Read ", len," instead of ", mem_segment_size, " bytes.");
-                        #endif
-                        return false;
-                    }
-                    p->set_from_bytes(mem);
-                    p->saved();        // Disable save flag when loading from memory.
-                }
-                else save_to_memory(p);
-            }
-            #ifdef DEBUG_PROPERTY
-            PRINT("[Properties]: Closing memory");
-            #endif
-            memory.end();
-            memory_opened = false;
-            return true;
-        }
-
-        void apply_setting(String key, String value)
-        {
-            #ifdef DEBUG_PROPERTY
-            PRINT("[Properties]: Applying ", value.c_str(), " to ", key.c_str());
-            #endif
-            bool found = false;
-            for(BaseProperty* p : properties)
-            {
-                if(strcmp(p->get_name(), key.c_str()) == 0)
-                {
-                    found = true;
-                    p->set_from_string(value);
-                    #ifdef DEBUG_PROPERTY
-                    PRINT("[Properties]: ", p->get_name(), " updated.");
-                    #endif
-                }
-            }
-            if(!found)
-            {
-                #ifdef DEBUG_PROPERTY
-                PRINT("[Properties] ", key.c_str(), " not found.");
-                #endif
-            }
-        }
-
-        void print_to(Print& sink)
-        {
-            sink.println("[Properties]: printing config.");
-            for(BaseProperty* p : properties) p->print_to(sink);
-        }
-
-        void save()
-        {
-            // Loop through all registered properties and check if they need saving.
-            for(BaseProperty* p : properties)
-            {
-                if(p->is_updated()) save_to_memory(p);
-            }
-            // Close memory if opened.
-            if(memory_opened)
-            {
-                #ifdef DEBUG_PROPERTY
-                PRINT("[Properties]: Closing memory");
-                #endif
-                memory.end();
-                memory_opened = false;
-            }
-        }
 
         BaseProperty* get_property(uint32_t i)
         {
@@ -342,52 +85,5 @@ class PropertyStore: public BaseStore
             else return properties[i];
         }
 };
-
-template<size_t SIZE>
-class TelemetryStore: public BaseStore
-{
-    private:
-        BaseProperty* variables[SIZE];
-    public:
-        TelemetryStore():
-            BaseStore(SIZE)
-        {}
-
-        TelemetryStore(BaseProperty* const (&vars)[SIZE]):
-            BaseStore(SIZE)
-        {
-            assign_properties(vars);
-        }
-
-        TelemetryStore<SIZE> operator=(const TelemetryStore<SIZE>& rhs)
-        {
-            assign_properties(rhs.variables);
-            return *this;
-        }
-
-        void assign_properties(BaseProperty* const (&vars)[SIZE])
-        {
-            for(uint32_t i = 0; i < SIZE; i++)
-            {
-                variables[i] = vars[i];
-            }
-        }
-
-        void print_to(Print& sink)
-        {
-            sink.println("[Variables]: printing config.");
-            for(BaseProperty* v : variables) v->print_to(sink);
-        }
-
-        BaseProperty* get_property(uint32_t i)
-        {
-            if(i >= SIZE) return nullptr;
-            else return variables[i];
-        }
-};
-
-#else
-#error "This architecture is not supported."
-#endif
 
 #endif